Tricky test 3

+++ main-3-5.cpp
@@ -2,7 +2,7 @@
 
 // The `extern` keyword tells the compiler that the count function exists,
 // but will be implemented somewhere else
-extern double sum_even(int array[], int n);
+extern double sum_even(double array[], int n);
 
 int main() {
     int array[5] = {4,5,6,7,8};
